+++ tests/conftest.py
@@ -17,6 +17,7 @@ def pytest_configure(config):  # noqa: ARG001
     os.environ["PYTHON_DOTENV_DISABLED"] = "1"
     os.environ.pop("GROQ_API_KEY", None)
     os.environ.pop("OPENROUTER_API_KEY", None)
+    os.environ.pop("COMPOSIO_API_KEY", None)
     os.environ.pop("SWARM_AGENT_MODEL", None)
     os.environ.pop("LANGFUSE_PUBLIC_KEY", None)
     os.environ.pop("LANGFUSE_SECRET_KEY", None)